sband.c: utility functions to work with the s-band transmitter The move of sband.c to hal_sband.c is temporary
