+++ ex2_hal/sband/hardware_interface/include/hal_sband.h
@@ -1,0 +1,100 @@
+/*
+ * Copyright (C) 2021-2022  University of Alberta
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+#ifndef HAL_SBAND_H
+#define HAL_SBAND_H
+
+#include <inttypes.h>
+#include "HL_reg_het.h"
+#include <sTransmitter.h>
+
+typedef enum {
+    COUNT = S_BUFFER_COUNT,
+    UNDERRUN = S_BUFFER_UNDERRUN,
+    OVERRUN = S_BUFFER_OVERRUN,
+} Buffer_Quantity;
+
+typedef enum {
+    PA_STATUS_DISABLE = 0,
+    PA_STATUS_ENABLE = 1
+} Sband_PowerAmplifier_Status_t;
+
+typedef enum {
+    PA_MODE_CONF = 0,
+    PA_MODE_SYNC = 1,
+    PA_MODE_DATA = 2,
+    PA_MODE_TEST = 3
+} Sband_Transmitter_Mode_t;
+
+typedef struct __attribute__((packed)) {
+    Sband_PowerAmplifier_Status_t status;
+    Sband_Transmitter_Mode_t mode;
+} Sband_PowerAmplifier;
+
+typedef struct __attribute__((packed)) {
+    uint8_t scrambler;
+    uint8_t filter;
+    uint8_t modulation;
+    uint8_t rate;
+    uint8_t bit_order;
+} Sband_Encoder;
+
+typedef struct __attribute__((packed)) {
+    float freq;
+    uint8_t PA_Power;
+    Sband_PowerAmplifier PA;
+    Sband_Encoder enc;
+} Sband_config;
+
+typedef struct __attribute__((packed)) {
+    uint8_t PWRGD;
+    uint8_t TXL;
+} Sband_Status;
+
+typedef struct __attribute__((packed)) {
+    uint8_t transmit;
+} Sband_TR;
+
+typedef struct __attribute__((packed)) {
+    uint16_t firmware;
+} Sband_FirmwareV;
+
+typedef struct __attribute__((packed)) {
+    uint16_t pointer[3];
+} Sband_Buffer;
+
+typedef struct __attribute__((packed)) {
+    Sband_Status status;
+    Sband_TR transmit;
+    Sband_FirmwareV firmware;
+    Sband_Buffer buffer;
+    Sband_Housekeeping HK;
+} Sband_Full_Status;
+
+STX_return HAL_S_getFreq(float *S_freq);
+STX_return HAL_S_getPAPower(uint8_t *S_PA_power);
+STX_return HAL_S_getControl(Sband_PowerAmplifier *S_PA);
+STX_return HAL_S_getEncoder(Sband_Encoder *S_Enc);
+STX_return HAL_S_getStatus(Sband_Status *S_status);
+STX_return HAL_S_getTR(Sband_TR *S_transmit);
+STX_return HAL_S_getHK(Sband_Housekeeping *S_hk);
+STX_return HAL_S_hk_convert_endianness(Sband_Housekeeping *S_hk);
+STX_return HAL_S_getBuffer(int quantity, Sband_Buffer *S_buffer);
+STX_return HAL_S_softResetFPGA(void);
+STX_return HAL_S_setFreq(float S_freq_new);
+STX_return HAL_S_setPAPower(uint8_t S_PA_Power_new);
+STX_return HAL_S_setControl(Sband_PowerAmplifier S_PA_new);
+STX_return HAL_S_setEncoder(Sband_Encoder S_enc_new);
+
+#endif /* HAL_SBAND_H */
