+++ ex2_hal/sband/hardware_interface/source/sband.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2015  University of Alberta
+ * Copyright (C) 2022 University of Alberta
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -11,197 +11,123 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  */
+
 /**
  * @file sband.c
- * @author Arash Yazdani
- * @date 2020-10-01
+ * @author Ron Unrau
+ * @date 2022-07-01
  */
-
-/*
- * When TRX connected, the stubbed blocks can be used for TRX = off situation.
- */
-#include "sband.h"
-
-#include <string.h>
 #include <FreeRTOS.h>
-#include <csp/csp_endian.h>
-#include <os_queue.h>
+#include <os_task.h>
+#include <sTransmitter.h>
+#include <sband.h>
+#include <logger/logger.h>
 
-#include "services.h"
+#define SBAND_SYNC_BYTES 8
+static uint16_t syncword[SBAND_SYNC_BYTES/sizeof(uint16_t)] = {0xdadb, 0x0dba, 0xbeee, 0xd00d};
 
-// For storing the set data
-static Sband_config S_config_reg;
-static Sband_Full_Status S_FS;
-
-STX_return HAL_S_getFreq(float *S_freq) {
-    STX_return status;
-#if SBAND_IS_STUBBED == 0
-    status = STX_getFrequency(&S_config_reg.freq);
-#else
-    status = IS_STUBBED_S;
-#endif
-    *S_freq = S_config_reg.freq;
-    return status;
-};
-
-STX_return HAL_S_getControl(Sband_PowerAmplifier *S_PA) {
-    STX_return status;
-#if SBAND_IS_STUBBED == 0
-    status = STX_getControl(&S_config_reg.PA.status, &S_config_reg.PA.mode);
-#else
-    status = IS_STUBBED_S;
-#endif
-    *S_PA = S_config_reg.PA;
-    return status;
-};
-
-STX_return HAL_S_getEncoder(Sband_Encoder *S_Enc) {
-    STX_return status;
-#if SBAND_IS_STUBBED == 0
-    status = STX_getEncoder(&S_config_reg.enc.bit_order, &S_config_reg.enc.scrambler, &S_config_reg.enc.filter,
-                            &S_config_reg.enc.modulation, &S_config_reg.enc.rate);
-#else
-    status = IS_STUBBED_S;
-#endif
-    *S_Enc = S_config_reg.enc;
-    return status;
-}
-
-STX_return HAL_S_getPAPower(uint8_t *S_PA_Power) {
-    STX_return status;
-#if SBAND_IS_STUBBED == 0
-    status = STX_getPaPower(&S_config_reg.PA_Power);
-#else
-    status = IS_STUBBED_S;
-#endif
-    *S_PA_Power = S_config_reg.PA_Power;
-    return status;
-};
-
-STX_return HAL_S_getFirmwareV(Sband_FirmwareV *S_firmwareV) {
-    STX_return status;
-#if SBAND_IS_STUBBED == 0
-    status = STX_getFirmwareV(&S_FS.firmware.firmware);
-#else
-    S_FS.firmware.firmware = 111;
-    status = IS_STUBBED_S;
-#endif
-    *S_firmwareV = S_FS.firmware;
-    return status;
-}
-
-STX_return HAL_S_getStatus(Sband_Status *S_status) {
-    STX_return status;
-#if SBAND_IS_STUBBED == 0
-    status = STX_getStatus(&S_FS.status.PWRGD, &S_FS.status.TXL);
-#else
-    S_FS.status.PWRGD = 1;
-    S_FS.status.TXL = 1;
-    status = IS_STUBBED_S;
-#endif
-    *S_status = S_FS.status;
-    return status;
-}
-
-STX_return HAL_S_getTR(Sband_TR *S_transmit) {
-    STX_return status;
-#if SBAND_IS_STUBBED == 0
-    status = STX_getTR(&S_FS.transmit.transmit);
-#else
-    S_FS.transmit.transmit = 1;
-    status = IS_STUBBED_S;
-#endif
-    *S_transmit = S_FS.transmit;
-    return status;
-}
-
-STX_return HAL_S_getHK(Sband_Housekeeping *S_hk) {
-    STX_return status;
-#if SBAND_IS_STUBBED == 0
-    status = STX_getHK(&S_FS.HK);
-#else
-    S_FS.HK.Output_Power = 26;
-    S_FS.HK.PA_Temp = 27.3;
-    S_FS.HK.Top_Temp = -2.8;
-    S_FS.HK.Bottom_Temp = 11.7;
-    S_FS.HK.Bat_Current = 95;
-    S_FS.HK.Bat_Voltage = 7.2;
-    S_FS.HK.PA_Current = 0.48;
-    S_FS.HK.PA_Voltage = 5.1;
-    status = IS_STUBBED_S;
-#endif
-    *S_hk = S_FS.HK;
-    return status;
-}
-
-STX_return HAL_S_hk_convert_endianness(Sband_Housekeeping *S_hk) {
-    S_hk->Output_Power = csp_htonflt(S_hk->Output_Power);
-    S_hk->PA_Temp = csp_htonflt(S_hk->PA_Temp);
-    S_hk->Top_Temp = csp_htonflt(S_hk->Top_Temp);
-    S_hk->Bottom_Temp = csp_htonflt(S_hk->Bottom_Temp);
-    S_hk->Bat_Current = csp_htonflt(S_hk->Bat_Current);
-    S_hk->Bat_Voltage = csp_htonflt(S_hk->Bat_Voltage);
-    S_hk->PA_Current = csp_htonflt(S_hk->PA_Current);
-    S_hk->PA_Voltage = csp_htonflt(S_hk->PA_Voltage);
-    S_hk->PA_Voltage = csp_htonflt(S_hk->PA_Voltage);
-    return S_SUCCESS;
-}
-
-/* The switch operation might be better implemented here than in EH */
-STX_return HAL_S_getBuffer(int quantity, Sband_Buffer *S_buffer) {
-    STX_return status;
-    /* Although there is no writing data, we can call a function like them*/
-#if SBAND_IS_STUBBED == 0
-    status = STX_getBuffer(quantity, &S_FS.buffer.pointer[quantity]);
-#else
-    S_FS.buffer.pointer[quantity] = quantity;
-    status = IS_STUBBED_S;
-#endif
-    *S_buffer = S_FS.buffer;
-    return status;
-}
-
-STX_return HAL_S_softResetFPGA(void) {
 #if SBAND_IS_STUBBED == 1
-    return IS_STUBBED_S;
+static int mode;
+#endif
+
+int sband_init() {
+    int ret;
+
+    STX_Enable();
+    vTaskDelay(2*ONE_SECOND);
+ 
+    if ((ret = STX_setControl(S_PA_DISABLE, S_CONF_MODE)) != S_SUCCESS) return ret;
+    ret = STX_setEncoder(S_BIT_ORDER_MSB, S_SCRAMBLER_DISABLE,
+                         S_FILTER_ENABLE, S_MOD_QPSK, S_RATE_FULL);
+    if (ret != S_SUCCESS) return ret;
+    if ((ret = STX_setFrequency(2228)) != S_SUCCESS) return ret;
+    if ((ret = STX_setFrequency(2228.0f)) != S_SUCCESS) return ret;
+    if ((ret = STX_setPaPower(30u)) != S_SUCCESS) return ret;
+
+    return (int) S_SUCCESS;
+}
+
+bool sband_enter_conf_mode() {
+    // Call this when you're done transmitting
+#if SBAND_IS_STUBBED == 1
+    STX_return ret = S_SUCCESS;
+    mode = S_CONF_MODE;
 #else
-    return STX_softResetFPGA();
+    Sband_PowerAmplifier pa = { .status = PA_STATUS_DISABLE, .mode = PA_MODE_CONF };
+    STX_return ret = HAL_S_setControl(pa);
+#endif
+    if (ret != S_SUCCESS) {
+        ex2_log("%s failed: %d", __FUNCTION__, ret);
+    }
+    return (ret == S_SUCCESS);
+}
+
+bool sband_enter_sync_mode() {
+    // Call this to get ready to transmit
+#if SBAND_IS_STUBBED == 1
+    STX_return ret = S_SUCCESS;
+    mode = S_SYNC_MODE;
+#else
+    Sband_PowerAmplifier pa = { .status = PA_STATUS_ENABLE, .mode = PA_MODE_SYNC };
+    STX_return ret = HAL_S_setControl(pa);
+#endif
+    if (ret != S_SUCCESS) {
+        ex2_log("%s failed: %d", __FUNCTION__, ret);
+    }
+
+    return (ret == S_SUCCESS);
+}
+
+bool sband_enter_data_mode() {
+    // Call this to start transmitting once the FIFO is full
+#if SBAND_IS_STUBBED == 1
+    STX_return ret = S_SUCCESS;
+    mode = S_DATA_MODE;
+#else
+    Sband_PowerAmplifier pa = { .status = PA_STATUS_ENABLE, .mode = PA_MODE_DATA };
+    STX_return ret = HAL_S_setControl(pa);
+#endif
+    if (ret != S_SUCCESS) {
+        ex2_log("%s failed: %d", __FUNCTION__, ret);
+    }
+    return (ret == S_SUCCESS);
+}
+
+void sband_sync() {
+    /* Send this before the first data byte and after every SBAND_SYNC_INTERVAL.
+     * Note that it won't actually be transmitted until you enter data mode.
+     * Also note that our S-Band SPI is 16-bits wide (hence the divide by 2).
+     */
+#if SBAND_IS_STUBBED == 1
+    ex2_log("S-band SYNC");
+#else
+    SPISbandTx(syncword, SBAND_SYNC_BYTES/sizeof(uint16_t));
 #endif
 }
 
-STX_return HAL_S_setFreq(float S_freq_new) {
-    S_config_reg.freq = S_freq_new;
+int sband_transmit_ready() {
+    /* This line goes high when there is less than 2560 bytes in the 20KB FIFO.
+     * The manual says to keep the FIFO as full as possible but avoid over-runs.
+     */
 #if SBAND_IS_STUBBED == 1
-    return IS_STUBBED_S;
+    return 1;
 #else
-    return STX_setFrequency(S_config_reg.freq);
+    return gioGetBit(hetPORT1, 25);
 #endif
 }
 
-STX_return HAL_S_setPAPower(uint8_t S_PA_Power_new) {
-    S_config_reg.PA_Power = S_PA_Power_new;
+bool sband_buffer_count(uint16_t *cnt) {
 #if SBAND_IS_STUBBED == 1
-    return IS_STUBBED_S;
+    STX_return ret = S_SUCCESS;
+    *cnt = 100;
 #else
-    return STX_setPaPower(S_config_reg.PA_Power);
+    STX_return ret = STX_getBuffer(S_BUFFER_COUNT, cnt);
 #endif
+    if (ret != S_SUCCESS) {
+        ex2_log("%s failed: %d", __FUNCTION__, ret);
+        *cnt = 0;
+    }
+    return (ret == S_SUCCESS);
 }
 
-STX_return HAL_S_setControl(Sband_PowerAmplifier S_PA_new) {
-    S_config_reg.PA = S_PA_new;
-#if SBAND_IS_STUBBED == 1
-    return IS_STUBBED_S;
-#else
-    return STX_setControl(S_config_reg.PA.status, S_config_reg.PA.mode);
-#endif
-}
-
-STX_return HAL_S_setEncoder(Sband_Encoder S_enc_new) {
-    S_config_reg.enc = S_enc_new;
-#if SBAND_IS_STUBBED == 1
-    return IS_STUBBED_S;
-#else
-    return STX_setEncoder(S_config_reg.enc.bit_order, S_config_reg.enc.scrambler, S_config_reg.enc.filter,
-                          S_config_reg.enc.modulation, S_config_reg.enc.rate);
-#endif
-}
+
