+++ ex2_services/Services/include/housekeeping/housekeeping_service.h
@@ -30,7 +30,7 @@
 #include "housekeeping_athena.h"
 #include "housekeeping_charon.h"
 #include "hyperion.h"
-#include "sband.h"
+#include "hal_sband.h"
 #include "uhf.h"
 #include "dfgm.h"
 #include "ns_payload.h"
