+++ ex2_system/source/test_sdr.c
@@ -53,7 +53,7 @@ void test_csp_send(void *arg) {
     csp_close(conn);
 }
 
-void test_sdr_send(void *arg) {
+void test_uhf_send(void *arg) {
     sdr_interface_data_t *ifdata = arg;
 
     ex2_log("starting SDR->UART send");
@@ -79,6 +79,38 @@ void test_sdr_send(void *arg) {
         }
         vPortFree(packet);
         vTaskDelay(5000);
+    }
+}
+
+void test_sband_send(void *arg) {
+    sdr_interface_data_t *ifdata = arg;
+    int len = 1024;
+    uint8_t *packet = (uint8_t *) pvPortMalloc(len);
+
+    ex2_log("starting S-BAND send");
+    vTaskDelay(1000);
+
+    while (1) {
+        /* Simulate the sending of 10MB files */
+        sdr_sband_tx_start(ifdata);
+
+        int i, count;
+        for (count=0; count < 10*1024; count++) {
+            for (i=0; i<len; i++) {
+                packet[i] = count & 0x0ff;
+            }
+            ex2_log("sband send %d, len %d", count, len);
+
+            int rc = sdr_sband_tx(ifdata, packet, len);
+            if (rc) {
+                ex2_log("sdr_sband_tx returns %d", rc);
+            }
+        }
+
+        ex2_log("sband rest %d", count/len);
+        sdr_sband_tx_stop(ifdata);
+        vTaskDelay(10000);
+        ex2_log("sband back at it!");
     }
 }
 
@@ -142,14 +174,17 @@ static void sdr_uhf_receive(void *conf, uint8_t *data, size_t len, void *unused)
     ex2_log("uhf->csp %d mismatches", mismatches);
 }
 
-void start_test_sdr(sdr_interface_data_t *ifdata) {
+void start_test_sdr(sdr_interface_data_t *uhf_ifdata, sdr_interface_data_t *sband_ifdata) {
     if (SDR_NO_CSP) {
-        ifdata->sdr_conf->rx_callback = sdr_uhf_receive;
-        ifdata->sdr_conf->rx_callback_data = NULL;
+        uhf_ifdata->sdr_conf->rx_callback = (sdr_rx_callback_t) sdr_uhf_receive;
+        uhf_ifdata->sdr_conf->rx_callback_data = NULL;
 
-        xTaskCreate(test_sdr_send, "sdr->uhf", TEST_STACK_SIZE, ifdata, 0, NULL);
-    } else {
+        xTaskCreate(test_uhf_send, "sdr->uhf", TEST_STACK_SIZE, uhf_ifdata, 0, NULL);
+    }
+    else {
         xTaskCreate(test_csp_receive, "uhf->csp", TEST_STACK_SIZE, NULL, 0, NULL);
         xTaskCreate(test_csp_send, "csp->uhf", TEST_STACK_SIZE, NULL, 0, NULL);
     }
+
+    xTaskCreate(test_sband_send, "sdr->sband", TEST_STACK_SIZE, sband_ifdata, 0, NULL);
 }
